moved classes from the duplicate question folders
